Unsupervised Learning -- Customer Segments
type: chore

The goal of this commit is to initialize a local repository for
a Udacity project. This one covers the Machine Learning Nanodegree
project about unsupervised learning. For those who are looking
to learn more about Machine Learning or Data Science, this
is a great project to use as a baseline as it eases users into
data preprocessing, feature engineering with PCA, clutstering
,and predicting using clusters.

In future iterations of this repo, I am hoping to clean up the code
and add in more ways to work with the data that will yeild
more accurate results and/or provide a flexible framework that
can work on multiple datasets.